- Tooltip sample updated
