+++ src/frontend-samples/tooltip-customize-sample/TooltipCustomize.scss
@@ -1,0 +1,16 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+.sample-options {
+  color: white;
+  padding: 8px;
+}
+
+.sample-options-3col {
+  display: grid;
+  row-gap: 8px;
+  column-gap: 8px;
+  grid-template-columns: 1fr 2fr 2fr;
+  align-items: center;
+}
